dry dry dry, re #9744
